BREAKING CHANGE: auto-init and function names
All setup/initialization functions have been renamed to `init`, and they
are now all called automatically when the module is loaded. If you were
using any of the affected modules, you can now remove the call to the
`setup` or similar function. If you need to call it for some reason
(i.e. to change some defaults), all of them are now called `init`.