New function util:getfile
(Courtesy of @hanche in the Elvish channel) On macOS, dragging a file
from the Finder into a Terminal window results in its path being
pasted. Unfortunately it gets pasted with escape characters which
Elvish does not always interpret correctly, and also with an extra
space at the end. The util:getfile function can be used instead of the
filename, and when you drag the file it captures and fixes the
filename.