Allow specifying arguments position in alias
Normally, any arguments you specify to the alias are appended at the
end of its definition. However, you can now specify where they should
appear by inserting the string {} (in Elvish you have to quote it like
this '{}' ), then the arguments are inserted at that place in the
command.

Thanks to alex`` for the idea!