Updated for 0.13 release
Modules bang-bang and dir have been updated for the new editor API
included in release 0.13 of Elvish. Do not use this version if you are
not running this version (or the latest changes in master leading up
to it). In this case please use the code from the pre-0.13 branch.