Initial Shell Integration support
The function iterm2:init sets up hooks to produce the appropriate
escape codes for iTerm2's shell integration. Note that this is at the
moment not ideal because it captures the prompt as part of the
command, but at the moment there is no way in Elvish to insert the
marker in the correct place.

Also reorganized and restructured the other functions in more readable
sections.