New iterm2:reportvar function
Instructs iTerm2 to report the value of a variable. Note that there is
no way for the shell to capture this report, it gets displayed in the
terminal itself. The value of the variable is returned base64-encoded,
so you need to pass it to the `base64 -d` command to see its value.