Fixed shell integration codes for correct command capture
Moved emission of the end-of-prompt code to the end of the prompt
function, leaving the others in the before/after-readline hooks. This
has the effect of capturing all the parts of the command and its
output correctly.

At the moment, proper capture of the command means modifying the
prompt function, which means =iterm2:init= needs to be called *after*
setting up your prompt. The original prompt function is saved in
=$iterm2:original-prompt-fn=, and used within the modified prompt
function to display it. We also modify the =$edit:before-readline= and
=$edit:after-readline= hooks to emit the corresponding escape
sequences in the correct places.

*Note:* at the moment Elvish does not have a mechanism for capturing
the exit code of the last command executed, therefore the FTCS
"command finished" marker is always emitted with an exit code of zero,
so the marker is always colored as if the command was successful.