Support empty titles better
Something changed in the behavior of edit:wordify (I think) which is now
resulting in an empty output instead of an empty string. This change
better handles this behavior instead of producing an error.

There are also a lot of whitespace changes due to indentation behavior
changes in my Emacs configuration.